fixed issue #2
http://code.google.com/p/blacktree-visor/issues/detail?id=2

The problem:
After sliding window off-screen there was a code block which positioned window to original place and called window orderOut to hide it. According to
http://developer.apple.com/documentation/Cocoa/Reference/ApplicationKit/Classes/NSWindow_Class/Reference/Reference.html#//apple_ref/occ/instm/NSWindow/orderOut:
"If the receiver is the key or main window, the NSWindow object immediately behind it is made key or main in its place", this was the case and caused next (non-visored) terminal window to become frontmost.

Solution:
I don't use orderOut to hide the window. I simply keep the window off-screen in the same state as it was left. Every call to showWindow first resets the position of off-screen window (counting with possibly new height) and starts slide animation to slide the window to top-screen-edge position. I've also added code which disables shadow of off-screen window to not interfere with top screen edge. It is also important to finish animation with integer values. Floating point values in translation cause artefacts on bottom window edge.